Detect short name of current environment.
Show an environment indicator for all environments and not just MSYS2.
To do this, use heuristics to shorten the name of the current Linux
distribution from /etc/os-release, or use the output of "uname -o" if
not available. Continue to use $MSYSTEM for MSYS2, evaluated
dynamically.
Update README.md for this change.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>