Support busybox for Windows, refactor
Fix the git repository detection heuristics to support `$PWD` on busybox
for Windows containing the `DRIVE:` prefix, which caused an infinite
loop on startup otherwise.

Try to detect busybox for Windows and possibly other Windows shells.

Refactor to use `local` for function variables, which is available on
all POSIX shells in current use, and add a main function.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>